Final Commit

+++ src/main/java/com/commonplant/umc/service/MemoService.java
@@ -3,6 +3,7 @@ package com.commonplant.umc.service;
 import com.commonplant.umc.config.exception.BadRequestException;
 import com.commonplant.umc.config.exception.ErrorResponseStatus;
 import com.commonplant.umc.domain.Memo;
+import com.commonplant.umc.domain.Place;
 import com.commonplant.umc.domain.Plant;
 import com.commonplant.umc.domain.User;
 import com.commonplant.umc.dto.memo.MemoRequest;
@@ -26,10 +27,18 @@ public class MemoService {
     private final MemoRepository memoRepository;
 
     private final FirebaseService firebaseService;
+    private final PlaceService placeService;
     private final PlantService plantService;
 
     @Transactional
     public Long addMemo(User user, Long plant, String content, MultipartFile file){
+
+        // TODO: 장소에 속해 있는 유저만 메모를 추가 가능하게 하기
+        Place place = plantService.getPlant(plant).getPlace();
+
+        if (!placeService.ExistUserInPlace(user, place)) {
+            throw new BadRequestException(ErrorResponseStatus.NOT_FOUND_USER_IN_PLACE);
+        }
 
         // imgUrl Setter
         String newCode = null;
@@ -42,7 +51,6 @@ public class MemoService {
             imgUrl = firebaseService.uploadFiles("commonPlant_memo_" + plantMemoIdx + "_" + newCode, file);
         }
 
-        // User user = userService.getUser(req.getWriter());
         Plant memoPlant = plantService.getPlant(plantMemoIdx);
 
         System.out.println(plantMemoIdx);
@@ -101,7 +109,14 @@ public class MemoService {
     }
 
     @Transactional
-    public MemoResponse.memoListRes getMemoList(Long plantIdx){
+    public MemoResponse.memoListRes getMemoList(Long plantIdx, User user){
+
+        // TODO: 장소에 속해 있는 유저만 메모 리스트를 조회 가능하게 하기
+        Place place = plantService.getPlant(plantIdx).getPlace();
+
+        if (!placeService.ExistUserInPlace(user, place)) {
+            throw new BadRequestException(ErrorResponseStatus.NOT_FOUND_USER_IN_PLACE);
+        }
 
         // Getter
         List<Memo> memoList = memoRepository.findAllByPlantIdxOrderByCreatedAtDesc(plantIdx);
@@ -143,47 +158,74 @@ public class MemoService {
     }
 
     @Transactional
-    public String updateMemo(Long memoIdx, User user, MemoRequest.updateMemo req, MultipartFile file){
-
-        // imgUrl Setter
-        String newCode = randomCode();
-        Long plantMemoIdx = req.getPlant();
-
-        String imgUrl = null;
-
-        if(file.getSize()>0){
-            imgUrl = firebaseService.uploadFiles("commonPlant_memo_" +
-                    plantMemoIdx + "_" + newCode,file);
-        }
-
-        Plant plant = plantService.getPlant(req.getPlant());
-        //System.out.println(plant.getPlantIdx());
+    public Long updateMemo(Long memoIdx, User user, String content, MultipartFile file){
 
         Memo memo = memoRepository.findByMemoIdx(memoIdx);
 
+        // TODO: 메모의 작성한 사람과 수정하려는 사람이 같은지 비교
+        User writer = memo.getWriter();
+
+        if (writer != user) {
+            throw new BadRequestException(ErrorResponseStatus.INVALID_USER_JWT);
+        }
+
+        // TODO: 식물의 애칭이 등록되어 있지 않거나 10자를 넘어갈 경우 예외처리
+        String memoContent = null;
+
+        if (content.length() == 0) {
+            throw new BadRequestException(ErrorResponseStatus.NO_MEMO_CONTENT);
+        } else if (content.length() <= 200) {
+            memoContent = content;
+        } else {
+            throw new BadRequestException(ErrorResponseStatus.LONG_MEMO_CONTENT);
+        }
+
+        // imgUrl Setter
+        // 이미지가 바뀌더라도 url은 똑같게!
+        String imgUrl = null;
+        String newCode = null;
+        Long plantMemoIdx = memoIdx;
+
+        // TODO: 4가지 경우 모두 고려하기
+        if (memo.getImgUrl() == null) {
+            if (file.getSize() > 0) {
+                newCode = randomCode();
+                imgUrl = firebaseService.uploadFiles("commonPlant_memo_" + plantMemoIdx + "_" + newCode, file);
+            }
+        } else {
+            if (file.getSize() > 0) {
+                imgUrl = memo.getImgUrl();
+            }
+        }
+
         memo.updateMemo(
-                plant,
-                user,
-                req.getContent(),
+                memoContent,
                 imgUrl
         );
 
         memoRepository.save(memo);
+//
+//        String updateMemoTest = " 식물 이름: " + memo.getPlant() +
+//                " 메모 수정자: " + memo.getWriter() + " 수정 내용: " + memo.getContent()
+//                + " 수정된 이미지 url: " + memo.getImgUrl()
+//                + " 메모가 수정된 날짜: " + memo.getCreatedAt();
+//
+//        System.out.println(updateMemoTest);
 
-        String updateMemoTest = " 식물 이름: " + memo.getPlant() +
-                " 메모 수정자: " + memo.getWriter() + " 수정 내용: " + memo.getContent()
-                + " 수정된 이미지 url: " + memo.getImgUrl()
-                + " 메모가 수정된 날짜: " + memo.getCreatedAt();
-
-        System.out.println(updateMemoTest);
-
-        return updateMemoTest;
+        return memo.getPlant().getPlantIdx();
     }
 
     @Transactional
-    public Long deleteMemo(Long memoIdx){
+    public Long deleteMemo(Long memoIdx, User user){
 
         Memo memo = memoRepository.findByMemoIdx(memoIdx);
+
+        // TODO: 메모의 작성한 사람과 수정하려는 사람이 같은지 비교
+        User writer = memo.getWriter();
+
+        if (writer != user) {
+            throw new BadRequestException(ErrorResponseStatus.INVALID_USER_JWT);
+        }
 
         memoRepository.deleteById(memoIdx);
